rename `h3utils` to `util`
